requirements.txt
azure-functions
urllib3
requests
